requirements.txt
autopep8==2.0.2
Pillow==9.4.0
pycodestyle==2.10.0
PySimpleGUI==4.60.4
tomli==2.0.1
